+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
-### 1.0.0 (2016-10-02)
-- NEW: Backup all sites to dropbox
-- NEW: Restore all sites from dropbox
+
+### v 1.1.0 — 2016-10-09
+- NEW: Better space management
+- NEW: Remove SITE or SITE_NAME backups by default
+
+### v 1.0.0 — 2016-10-02
+- NEW: Backup all sites to Dropbox
+- NEW: Restore all sites from Dropbox